Replace Assembly.Location and publish single-file
This commit replaces any instances of `Assembly.Location` in
XIVLauncher.Core and XIVLauncher.Common with the appropriate
alternative, usually `AppContext.BaseDirectory`. Any such uses
of `Assembly.Location` would result in `string.Empty` when used
in the context of a single-file executable.

This squashed version of the prior patch no longer sets the project
to be self-contained, as it just hampers the build process outside of
publishing.

Ideally, any person building this project for own use now runs
`dotnet publish -c Release -r <linux-x64|osx-x64|win10-x64` to publish
the project.